# packages/api-client/openapi.json
{"openapi":"3.0.1","info":{"title":"HeatRoute API","description":"LCT 2026 district-heating route calculation backend","version":"0.1.0-java"},"servers":[{"url":"http://127.0.0.1:8000","description":"Generated server url"}],"paths":{"/api/v1/official/imports":{"post":{"tags":["official-import"],"summary":"Validate and persist an official LCT GeoJSON collection in PostGIS","operationId":"createOfficialImport","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialImportView"}}}}}}},"/api/v1/official/imports/{importId}/jobs/topology":{"post":{"tags":["official-jobs"],"summary":"Queue a durable topology-analysis job","operationId":"createTopologyJob","parameters":[{"name":"importId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialJobView"}}}}}}},"/api/v1/official/imports/inspect":{"post":{"tags":["official-import"],"summary":"Stream and validate the official LCT GeoJSON input contract","operationId":"inspectOfficialGeoJson","requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialInputReport"}}}}}}},"/api/v1/official/jobs/{jobId}":{"get":{"tags":["official-jobs"],"summary":"Read durable official job state and result","operationId":"getOfficialJob","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialJobView"}}}}}},"delete":{"tags":["official-jobs"],"summary":"Request cooperative cancellation","operationId":"cancelOfficialJob","parameters":[{"name":"jobId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialJobView"}}}}}}},"/api/v1/official/imports/{importId}":{"get":{"tags":["official-import"],"summary":"Read a durable official import report","operationId":"getOfficialImport","parameters":[{"name":"importId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficialImportView"}}}}}}},"/api/v1/official/imports/{importId}/topology":{"get":{"tags":["official-import"],"summary":"Validate upstream topology and create deterministic automatic tie-in candidates","operationId":"analyzeOfficialTopology","parameters":[{"name":"importId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TopologyAnalysis"}}}}}}},"/api/v1/health/ready":{"get":{"tags":["health"],"summary":"PostGIS readiness","operationId":"ready","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ReadinessResponse"}}}}}}},"/api/v1/health/live":{"get":{"tags":["health"],"summary":"Process liveness","operationId":"live","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}}},"components":{"schemas":{"OfficialImportView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"state":{"type":"string"},"originalFilename":{"type":"string"},"inputSizeBytes":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"report":{"$ref":"#/components/schemas/OfficialInputReport"}}},"OfficialInputError":{"type":"object","properties":{"code":{"type":"string"},"feature_index":{"type":"integer","format":"int64","writeOnly":true},"feature_id":{"type":"string","writeOnly":true},"field":{"type":"string"},"message":{"type":"string"},"featureIndex":{"type":"integer","format":"int64"},"featureId":{"type":"string"}}},"OfficialInputReport":{"type":"object","properties":{"contractVersion":{"type":"string"},"sha256":{"type":"string"},"featureCount":{"type":"integer","format":"int64"},"featureCounts":{"type":"object","additionalProperties":{"type":"integer","format":"int64"}},"errors":{"type":"array","items":{"$ref":"#/components/schemas/OfficialInputError"}},"valid":{"type":"boolean"}}},"JsonNode":{"type":"object"},"OfficialJobView":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"importId":{"type":"string","format":"uuid"},"jobType":{"type":"string"},"state":{"type":"string"},"phase":{"type":"string"},"progressCurrent":{"type":"integer","format":"int64"},"progressTotal":{"type":"integer","format":"int64"},"attempt":{"type":"integer","format":"int32"},"cancellationRequested":{"type":"boolean"},"result":{"$ref":"#/components/schemas/JsonNode"},"errorCode":{"type":"string"},"errorMessage":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"completedAt":{"type":"string","format":"date-time"}}},"TieInCandidate":{"type":"object","properties":{"connectionPointId":{"type":"string"},"targetId":{"type":"string"},"targetType":{"type":"string"},"distanceM":{"type":"number"},"newChamberRequired":{"type":"boolean"}}},"TopologyAnalysis":{"type":"object","properties":{"sourceCount":{"type":"integer","format":"int32"},"networkSegmentCount":{"type":"integer","format":"int32"},"chamberCount":{"type":"integer","format":"int32"},"issues":{"type":"array","items":{"$ref":"#/components/schemas/TopologyIssue"}},"tieInCandidates":{"type":"array","items":{"$ref":"#/components/schemas/TieInCandidate"}},"valid":{"type":"boolean"}}},"TopologyIssue":{"type":"object","properties":{"code":{"type":"string"},"featureId":{"type":"string"},"message":{"type":"string"}}},"DependencyStatus":{"type":"object","properties":{"status":{"type":"string"},"detail":{"type":"string"}}},"ReadinessResponse":{"type":"object","properties":{"status":{"type":"string"},"checks":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/DependencyStatus"}}}}}}}